requirements.txt
appdirs==1.4.3
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage-blob==0.37.1
azure-storage-common==0.37.1
azure-storage-nspkg==3.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
distlib==0.3.0
filelock==3.0.12
futures==3.1.1
idna==2.9
importlib-metadata==1.5.0
pycparser==2.20
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
virtualenv==20.0.12
zipp==3.1.0
